feat: distrubuted to folders and cleaned code
