+++ client/src/Molecules/NavButton/NavButtonWrapper.js
@@ -6,6 +6,5 @@ const NavButtonWrapper = styled.div`
   gap: 1rem;
   align-items: center;
   padding: 0 1rem;
-  background: ${({ theme }) => theme.grey00};
 `;
 export default NavButtonWrapper;
