+++ client/src/Molecules/NavCornerMenu/NavCornerMenuStyled.js
@@ -3,7 +3,6 @@ import styled from 'styled-components';
 const NavCornerMenuStyled = styled.div`
   text-align: left;
   width: 100%;
-  background: ${({ theme }) => theme.grey00};
   &.close {
     transform: scale(1, 0);
   }
