--- client/src/Organisms/Navbar/ButtonsWrapperStyled.js
@@ -1,11 +1,0 @@
-import styled from 'styled-components';
-
-const ButtonsWrapperStyled = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  min-width: 350px;
-  gap: 1rem;
-`;
-
-export default ButtonsWrapperStyled;
