--- client/src/Organisms/Navbar/NavMenuStyled.js
@@ -1,15 +1,0 @@
-import styled from 'styled-components';
-
-const NavMenuStyled = styled.div`
-  /* display: flex; */
-  display: block;
-  position: relative;
-  top: 2rem;
-  border: 5px dotted grey;
-  /* grid-area: 1 / 4 / 2 / 5; */
-  /* gap: 2rem; */
-  /* position: absolute; */
-  /* right: 10rem; */
-`;
-
-export default NavMenuStyled;
